move events to constants
