+++ remme/tp/atomic_swap.py
@@ -28,7 +28,7 @@ from remme.clients.account import AccountClient
 from remme.tp.account import AccountHandler, get_account_by_address
 from remme.shared.singleton import singleton
 from remme.ws.basic import EMIT_EVENT
-from remme.ws.events import Events
+from remme.ws.constants import Events
 
 LOGGER = logging.getLogger(__name__)
